Extents

Documentation for Extents.

Extents.Extent

Extents.buffer

Extents.extent

Extents.intersection

Extents.intersects

Extents.union

Extents.Extent — Type

Extent

Extent(; kw...)
Extent(bounds::NamedTuple)

A wrapper for a NamedTuple of tuples holding the lower and upper bounds for each dimension of the object.

keys(extent) will return the dimension name Symbols, in the order the dimensions are used in the object.

values(extent) will return a tuple of tuples: (lowerbound, upperbound) for each dimension.

Examples

julia> ext = Extent(X = (1.0, 2.0), Y = (3.0, 4.0))
Extent(X = (1.0, 2.0), Y = (3.0, 4.0))

julia> keys(ext)
(:X, :Y)

julia> values(ext)
((1.0, 2.0), (3.0, 4.0))

Extents.buffer — Method

buffer(ext::Extent, buff::NamedTuple)

buffer Extent by corresponding name-pair values supplied in buff NamedTuple.

Examples

julia> ext = Extent(X = (1.0, 2.0), Y = (3.0, 4.0))
Extent(X = (1.0, 2.0), Y = (3.0, 4.0))
julia> ext_buffered = Extents.buffer(ext, (X=1, Y=3))
Extent(X = (0.0, 3.0), Y = (0.0, 7.0))

Extents.extent — Function

extent(x)

Returns an Extent, holding the bounds for each dimension of the object.

Extents.intersection — Method

intersection(ext1::Extent, ext2::Extent; strict=false)

Get the intersection of two extents as another Extent, e.g. the area covered by the shared dimensions for both extents.

If there is no intersection for any shared dimension, nothing will be returned. When strict=true, any unshared dimensions cause the function to return nothing.

Extents.intersects — Method

intersects(ext1::Extent, ext2::Extent; strict=false)

Check if two Extent objects intersect.

Returns true if the extents of all common dimensions share some values including just the edge values of their range.

Dimensions that are not shared are ignored by default, with strict=false. When strict=true, any unshared dimensions cause the function to return false.

The order of dimensions is ignored in both cases.

If there are no common dimensions, false is returned.

Extents.union — Method

union(ext1::Extent, ext2::Extent; strict=false)

Get the union of two extents, e.g. the combined extent of both objects for all dimensions.

Dimensions that are not shared are ignored by default, with strict=false. When strict=true, any unshared dimensions cause the function to return nothing.